document-to-json-mvp/backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
pdfplumber==0.11.4
python-docx==1.1.0
pytesseract==0.3.13
Pillow==10.4.0
python-multipart==0.0.9
orjson==3.10.7
SQLModel==0.0.22
